Compiled code with changes to 2D byte array so no dynamic memory issues. Changes to folder formatting, needs better include access so no duplicate files

+++ dobotLibrary/counter.h
@@ -2,15 +2,13 @@
 #ifndef COUNTER_H
 #define COUNTER_H
 
-#include <iostream>
-
 class Counter{
 
     public:
         Counter();
-        uint8_t whiteValue;
-        uint8_t redValue;
-        uint8_t blackValue;
+        int whiteValue;
+        int redValue;
+        int blackValue;
 
 };
